+++ src/webserver/database/migrations.rs
@@ -1,5 +1,5 @@
+use super::error_highlighting::display_db_error;
 use super::Database;
-use crate::webserver::database::highlight_sql_error;
 use crate::MIGRATIONS_DIR;
 use anyhow;
 use anyhow::Context;
@@ -35,7 +35,7 @@ pub async fn apply(config: &crate::app_config::AppConfig, db: &Database) -> anyh
         match err {
             MigrateError::Execute(n, source) => {
                 let migration = migrator.iter().find(|&m| m.version == n).unwrap();
-                highlight_sql_error("Error in the SQL migration", &migration.sql, source).context(
+                display_db_error("Error in the SQL migration", &migration.sql, source).context(
                     format!("Failed to apply migration {}", DisplayMigration(migration)),
                 )
             }